added roofs and lights
I added roofs and lights to the kitchen and living room space - I also started adding some placeholder colors as it was very hard to distinguish between objects in dimmer lighting when they were all the same material. The space still looks very drab and sad for some reason, but I think that will change when the sun rises.